*: move Output to config.py
We want to expose this to users, which means we should have it in files
they import, not in backend/base/core.py.
