+++ libqtile/core/manager.py
@@ -21,12 +21,22 @@ from typing import TYPE_CHECKING
 import libqtile
 from libqtile import bar, hook, ipc, utils
 from libqtile.backend import base
-from libqtile.backend.base.core import Output
 from libqtile.command import interface
 from libqtile.command.base import CommandError, CommandException, CommandObject, expose_command
 from libqtile.command.client import InteractiveCommandClient
 from libqtile.command.interface import IPCCommandServer, QtileCommandInterface
-from libqtile.config import Click, Drag, Key, KeyChord, Match, Mouse, Rule, Screen, ScreenRect
+from libqtile.config import (
+    Click,
+    Drag,
+    Key,
+    KeyChord,
+    Match,
+    Mouse,
+    Output,
+    Rule,
+    Screen,
+    ScreenRect,
+)
 from libqtile.config import ScratchPad as ScratchPadConfig
 from libqtile.core.lifecycle import lifecycle
 from libqtile.core.loop import LoopContext
